Remove prompt from code blocks so user can just copy/paste
The code blocks have the prompt characters in them which means the user can't just use the copy button to copy to their terminal. 
In this suggested patch I've removed the prompts and put a comment character in front of some other notes.
Probably a style thing. You can ignore of course!